Move MerkleTreeId to types to avoid dependency on world-state from aztec-rpc
