+++ yarn-project/aztec-rpc/src/database/memory_db.ts
@@ -5,7 +5,7 @@ import { TxAuxDataDao } from './tx_aux_data_dao.js';
 import { TxDao } from './tx_dao.js';
 import { AztecAddress } from '@aztec/foundation/aztec-address';
 import { Fr, Point } from '@aztec/foundation/fields';
-import { MerkleTreeId } from '@aztec/world-state';
+import { MerkleTreeId } from '@aztec/types';
 
 export class MemoryDB extends MemoryContractDatabase implements Database {
   private txTable: TxDao[] = [];
