+++ yarn-project/sequencer-client/src/block_builder/solo_block_builder.ts
@@ -24,8 +24,8 @@ import {
 } from '@aztec/circuits.js';
 import { computeContractLeaf } from '@aztec/circuits.js/abis';
 import { LeafData, SiblingPath } from '@aztec/merkle-tree';
-import { ContractData, L2Block, PublicDataWrite } from '@aztec/types';
-import { MerkleTreeId, MerkleTreeOperations } from '@aztec/world-state';
+import { MerkleTreeId, ContractData, L2Block, PublicDataWrite } from '@aztec/types';
+import { MerkleTreeOperations } from '@aztec/world-state';
 import chunk from 'lodash.chunk';
 import flatMap from 'lodash.flatmap';
 import times from 'lodash.times';
