+++ yarn-project/sequencer-client/src/sequencer/utils.ts
@@ -1,5 +1,6 @@
 import { CombinedHistoricTreeRoots, Fr, PrivateHistoricTreeRoots } from '@aztec/circuits.js';
-import { MerkleTreeId, MerkleTreeOperations } from '@aztec/world-state';
+import { MerkleTreeId } from '@aztec/types';
+import { MerkleTreeOperations } from '@aztec/world-state';
 
 /**
  * Fetches the private, nullifier, and contract tree roots from a given db and assembles a CombinedHistoricTreeRoots object.
