+++ yarn-project/types/src/index.ts
@@ -1,12 +1,13 @@
-export * from './tx.js';
-export * from './tx_hash.js';
 export * from './constants.js';
-export * from './l2_block_downloader/index.js';
 export * from './contract_data.js';
 export * from './l1_addresses.js';
 export * from './l2_block.js';
 export * from './l2_block_context.js';
+export * from './l2_block_downloader/index.js';
 export * from './l2_block_source.js';
+export * from './merkle_tree_id.js';
 export * from './public_data_write.js';
+export * from './tx.js';
+export * from './tx_hash.js';
 export * from './unverified_data.js';
 export * from './unverified_data_source.js';
