+++ yarn-project/world-state/src/synchroniser/server_world_state_synchroniser.test.ts
@@ -3,9 +3,9 @@ import { AppendOnlyTreeSnapshot, NUMBER_OF_L1_L2_MESSAGES_PER_ROLLUP } from '@az
 import { fr } from '@aztec/circuits.js/factories';
 
 import { INITIAL_LEAF, Pedersen, SiblingPath } from '@aztec/merkle-tree';
-import { ContractData, L2Block, L2BlockSource, PublicDataWrite } from '@aztec/types';
+import { ContractData, L2Block, L2BlockSource, MerkleTreeId, PublicDataWrite } from '@aztec/types';
 import { jest } from '@jest/globals';
-import { MerkleTreeDb, MerkleTreeId } from '../index.js';
+import { MerkleTreeDb } from '../index.js';
 import { ServerWorldStateSynchroniser } from './server_world_state_synchroniser.js';
 import { WorldStateRunningState } from './world_state_synchroniser.js';
 import { Fr } from '@aztec/foundation/fields';
