+++ yarn-project/world-state/src/world-state-db/merkle_trees.ts
@@ -29,11 +29,11 @@ import {
   INITIAL_NULLIFIER_TREE_SIZE,
   IndexedTreeId,
   MerkleTreeDb,
-  MerkleTreeId,
   MerkleTreeOperations,
   PublicTreeId,
   TreeInfo,
 } from './index.js';
+import { MerkleTreeId } from '@aztec/types';
 import { SerialQueue } from '@aztec/foundation/fifo';
 
 /**
